ASL/requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
boto3==1.17.23
botocore==1.20.112
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
comtypes==1.3.0
contourpy==1.1.1
cryptography==42.0.5
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.9.0
cvzone==1.5.6
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==3.0.5
django-widget-tweaks==1.4.8
easydict==1.12
easyocr==1.7.1
enchant==0.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
filelock==3.13.1
flasgger==0.9.5
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.10
Flask-Uploads==0.2.1
flatbuffers==23.5.26
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.0.0
greenlet==3.0.3
grpcio==1.60.0
gTTS==2.5.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.4.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
imutils==0.5.4
ipykernel==6.29.3
ipython==8.12.3
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==2.10.1
jmespath==0.10.0
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.0
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter-server==2.13.0
jupyter-server-terminals==0.5.2
jupyterlab==4.1.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.3
keras==2.13.1
Keras-Applications==1.0.8
keras-vggface==0.6
kiwisolver==1.3.1
lazy-loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
lxml==5.1.0
Markdown==3.5.1
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow3-annotations==1.1.0
matplotlib==3.4.1
matplotlib-inline==0.1.6
mediapipe==0.10.9
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.4
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
notebook==7.1.1
notebook-shim==0.2.4
np-utils==0.6.0
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.17.1
opencv-contrib-python==4.9.0.80
opencv-python==4.5.5.64
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
oscrypto==1.3.0
overrides==7.7.0
packaging==23.2
pandas==1.2.4
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.1
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
playsound==1.2.2
pooch==1.8.1
premailer==3.10.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pybind11==2.11.1
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.7.0
pydantic-core==2.18.1
pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.2.0
pypiwin32==223
pypng==0.20220715.0
pyreadline3==3.4.1
pyrsistent==0.20.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2021.1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
referencing==0.33.0
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-aws4auth==1.1.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.3.7
scikit-image==0.21.0
scikit-learn==0.24.1
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.2
shapely==2.0.3
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.1
SQLAlchemy==2.0.27
sqlparse==0.3.1
stack-data==0.6.3
starlette==0.37.2
statsd==3.2.1
svglib==1.5.1
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tkintertable==1.3.3
tomli==2.0.1
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
ttkthemes==3.2.2
twilio==8.13.0
types-python-dateutil==2.8.19.20240311
typing==3.6.4
typing-extensions==4.10.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.24
uri-template==1.3.0
uritools==4.0.2
urllib3==1.26.18
utils==1.0.2
wavio==0.0.8
wcwidth==0.2.13
webcam==1.34
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==0.15.2
wordcloud==1.9.3
wrapt==1.16.0
WTForms==2.2.1
xgboost==2.0.3
xhtml2pdf==0.2.15
yagmail==0.15.293
yarl==1.9.4
zipp==3.17.0
